Refactor for version 1.8
