+++ DeservingCompanions/_Module/SubModule.xml
@@ -2,7 +2,7 @@
 <Module>
   <Name value="Deserving Companions" />
   <Id value="$modulename$" />
-  <Version value="v1.2.3" />
+  <Version value="v1.3.0" />
   <Official value="false" />
   <DefaultModule value="false" />
   <SingleplayerModule value="true" />
